requirements.txt
python-dotenv
apache-airflow-providers-snowflake